pyproject.toml
[tool.poetry]
name = "switchbot-dashboard"
version = "0.1.0"
description = "Dashboard for Switchbot using InfluxDB and Grafana"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
influxdb-client = "^1.36.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.11.4"
pytest = "^7.2.0"
types-requests = "^2.28.11.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
